Crate test_util

Re-exports

pub use crate::encryption::*;

Modules

encryption

kv_generator

logging

macros

runner
Useful functions for custom test framework.

security

Macros

retry

Structs

KvGenerator
A random generator of kv.

Constants

MIN_LOCAL_PORT
Linux by default use [32768, 61000] for local port.

Statics

INITIAL_PORT

Traits

TestHook
Hooks for customize tests procedure.

Functions

alloc_port
Allocates a port for testing purpose.

clear_failpoints
During panic, due to drop order, failpoints will not be cleared before tests exit. If tests wait for a sleep failpoint, the whole tests will hang. So we need a method to clear failpoints explicitly besides teardown.

init_log_for_test

new_channel_cred

new_security_cfg

run_failpoint_tests
A runner function for running failpoint tests.

run_test_with_hook
Connects std tests and custom test framework.

run_tests
A runner function for running general tests.

setup_for_ci