Module test_util::runner

Useful functions for custom test framework.

See https://doc.rust-lang.org/unstable-book/language-features/custom-test-frameworks.html.

Structs

CaseLifeWatcher

FailpointHook

Nope
A special TestHook that does nothing.

Constants

FS

Traits

TestHook
Hooks for customize tests procedure.

Functions

clear_failpoints
During panic, due to drop order, failpoints will not be cleared before tests exit. If tests wait for a sleep failpoint, the whole tests will hang. So we need a method to clear failpoints explicitly besides teardown.

run_failpoint_tests
A runner function for running failpoint tests.

run_test_with_hook
Connects std tests and custom test framework.

run_tests
A runner function for running general tests.